create new repo and projects
The original Harbin Repairs repo has some issues that
make it difficult to work with: very long filenames,
spaces in filenames, etc.

This repo contains my modified versions of the adapter
PCBs, but with changes to make the files easier to work
with, and to make the projects more self-contained.