move front-panel USB2 connector
In the previous version, the part of the PCB where the
USB2 connector was located bumped up against a capacitor
on the motherboard. In the revised version, the
USB2 connector is moved to the "bottom" of the adapter
PCB, so the cutout can extended to accommodate the
capacitor.

I think the adapter will hang off the edge of the
motherboard slightly towards the "bottom" of
the adapter PCB, but this edge of the motherboard
aligns with the front of the case, and I have never
seen a case that fits the motherboard flush with
the front of the case.