redesigned to move the connectors further apart
In the previous version, the pins of the 2x3 2.54mm
connector bumped up against the shroud of the 2x3 2mm
connector on the motherboard. (Not so much that the
adapter didn't work, but still not ideal.)

The revised version reorients the 2x3 2.54mm header and
moves it a bit further from the 2mm connector.